chore: replace `useModalModelDelete` with `DeleteModelModal`
